feat: добавлена обработка кейса, когда руками введен неправильный айди
