+++ src/index.tsx
@@ -4,6 +4,16 @@ import { App } from './components/app/app';
 import { Provider } from 'react-redux';
 import { store } from './store';
 import { ErrorMessage } from './components/error-message/error-message';
+import { fetchOffers } from './store/api-actions';
+import { checkAuth } from './store/api-actions';
+
+store.dispatch(checkAuth());
+store.dispatch(fetchOffers()).then(() => {
+  const offers = store.getState().offers;
+  if (offers) {
+    localStorage.setItem('offers', JSON.stringify(offers));
+  }
+});
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
