+++ src/pages/favorites/favorites.tsx
@@ -4,19 +4,14 @@ import { useAppSelector } from '../../hooks/useAppSelector/useAppSelector';
 import { Header } from '../../components/header/header';
 import { AppRoute } from '../../const';
 import { FavoritesList } from '../../components/favorites-list/favorites-list';
-import { NotFound } from '../404/404';
 import * as selectors from '../../store/selectors';
 import { createSelector } from '@reduxjs/toolkit';
 import { OfferType } from '../../components/types/offer';
 
 export function Favorites() {
-  const offers = useAppSelector(selectors.offers);
+
   const filteredOffers = createSelector(selectors.offers, (state) => state?.filter((offer) => offer.isFavorite));
   const favorites = useAppSelector(filteredOffers) as OfferType[];
-
-  if (offers === null) {
-    return (<NotFound />);
-  }
 
   return (
     <div className="page">
